\name{sobSenAn-package}
\alias{sobSenAn-package}
\alias{sobSenAn}
\docType{package}
\title{
\packageTitle{sobSenAn}
}
\description{
\packageDescription{sobSenAn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sobSenAn}
\packageIndices{sobSenAn}
}
\author{
\packageAuthor{sobSenAn}

Maintainer: \packageMaintainer{sobSenAn}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
