\name{sobSenAn-package}
\alias{sobSenAn-package}
\alias{sobSenAn}
\docType{package}
\title{
\packageTitle{sobSenAn}
}
\description{
\packageDescription{sobSenAn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sobSenAn}
\packageIndices{sobSenAn}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sobSenAn}

Maintainer: \packageMaintainer{sobSenAn}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
